Project folder restructuring
