--- src/Components/Group/Users/Users.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Users() {
-  return <>Usersista päivää</>;
-}
-
-export default Users;
