+++ src/Components/Main Page/New/New.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
-import validate from '../../Functions/validate';
+import validate from '../../../Functions/validate';
 import './style.css';
 
 function New(appProps) {
